#7568 Fix issue with data dir
